Fixed bug that caused the app to crash due to lack of permissions, by moving the permission request to main activity. Fix a bug that caused app to crash by editing when total meters walked updates. App still crashes if landscape mode

+++ kotlin-anko-template/app/src/main/java/dk/cphbusiness/template/Objective.kt
@@ -20,7 +20,7 @@ class DiscoverObjective(val latitude : Double, val longitude : Double, objName :
 
 }
 
-class SprintObjective(val timeLimit : Int, objName : String, objGoal : String, isCompleted : Boolean = false) : Objective(objName, objGoal, isCompleted) {
+class SprintObjective(val timeLimit : Long, objName : String, objGoal : String, isCompleted : Boolean = false) : Objective(objName, objGoal, isCompleted) {
 
     // Constructor
     init {
